api shipper

+++ bookshop/bookshop/src/main/java/com/example/bookshop/repository/OrderRepository.java
@@ -38,7 +38,7 @@ public interface OrderRepository extends JpaRepository<Order, Integer> {
 
     List<Order> findByStatusInAndShipperIsNull(List<String> pending);
 
-    List<Order> findByShipperIdAndStatus(Long shipperId, String status);
+    List<Order> findByShipperIdAndStatus(Integer shipperId, String status);
 //    Collection<Object> findAllByUserIdAndStatus(Integer userId, String status);
 //    Optional<Order> findByUserId(Integer userId);
 //    Optional<Order> findActiveOrderByUserId(Integer userId);
